[DBMongo]+::insertRequest()
