+++ src/DBMongo.php
@@ -7,6 +7,7 @@ use Longman\TelegramBot\Entities\Chat;
 use Longman\TelegramBot\Entities\ChosenInlineResult;
 use Longman\TelegramBot\Entities\InlineQuery;
 use Longman\TelegramBot\Entities\Message;
+use Longman\TelegramBot\Entities\Update;
 use Longman\TelegramBot\Entities\User;
 use Longman\TelegramBot\Exception\TelegramException;
 use MongoDB\Client;
@@ -278,6 +279,119 @@ class DBMongo extends DBBase
         ]);
 
         return $this->getInsertOneResult($insertOneResult);
+    }
+
+    public function insertRequest(Update $update)
+    {
+        if (!$this->isDbConnected()) {
+            return false;
+        }
+
+        $update_id   = $update->getUpdateId();
+        $update_type = $update->getUpdateType();
+
+        // @todo Make this simpler: if ($message = $update->getMessage()) ...
+        if ($update_type === 'message') {
+            $message = $update->getMessage();
+
+            if ($this->insertMessageRequest($message)) {
+                $message_id = $message->getMessageId();
+                $chat_id    = $message->getChat()->getId();
+
+                return $this->insertTelegramUpdate(
+                    $update_id,
+                    $chat_id,
+                    $message_id
+                );
+            }
+        } elseif ($update_type === 'edited_message') {
+            $edited_message = $update->getEditedMessage();
+
+            if ($edited_message_local_id = $this->insertEditedMessageRequest($edited_message)) {
+                $chat_id                 = $edited_message->getChat()->getId();
+
+                return $this->insertTelegramUpdate(
+                    $update_id,
+                    $chat_id,
+                    null,
+                    null,
+                    null,
+                    null,
+                    $edited_message_local_id
+                );
+            }
+        } elseif ($update_type === 'channel_post') {
+            $channel_post = $update->getChannelPost();
+
+            if ($this->insertMessageRequest($channel_post)) {
+                $message_id = $channel_post->getMessageId();
+                $chat_id    = $channel_post->getChat()->getId();
+
+                return $this->insertTelegramUpdate(
+                    $update_id,
+                    $chat_id,
+                    $message_id
+                );
+            }
+        } elseif ($update_type === 'edited_channel_post') {
+            $edited_channel_post = $update->getEditedChannelPost();
+
+            if ($edited_channel_post_local_id = $this->insertEditedMessageRequest($edited_channel_post)) {
+                $chat_id                      = $edited_channel_post->getChat()->getId();
+
+                return $this->insertTelegramUpdate(
+                    $update_id,
+                    $chat_id,
+                    null,
+                    null,
+                    null,
+                    null,
+                    $edited_channel_post_local_id
+                );
+            }
+        } elseif ($update_type === 'inline_query') {
+            $inline_query = $update->getInlineQuery();
+
+            if ($this->insertInlineQueryRequest($inline_query)) {
+                $inline_query_id = $inline_query->getId();
+
+                return $this->insertTelegramUpdate(
+                    $update_id,
+                    null,
+                    null,
+                    $inline_query_id
+                );
+            }
+        } elseif ($update_type === 'chosen_inline_result') {
+            $chosen_inline_result = $update->getChosenInlineResult();
+
+            if ($chosen_inline_result_local_id = $this->insertChosenInlineResultRequest($chosen_inline_result)) {
+                return $this->insertTelegramUpdate(
+                    $update_id,
+                    null,
+                    null,
+                    null,
+                    $chosen_inline_result_local_id
+                );
+            }
+        } elseif ($update_type === 'callback_query') {
+            $callback_query = $update->getCallbackQuery();
+
+            if ($this->insertCallbackQueryRequest($callback_query)) {
+                $callback_query_id = $callback_query->getId();
+
+                return $this->insertTelegramUpdate(
+                    $update_id,
+                    null,
+                    null,
+                    null,
+                    null,
+                    $callback_query_id
+                );
+            }
+        }
+
+        return false;
     }
 
     /**
